requirements.txt
selenium==4.15.1
Flask~=3.1.1
flask-cors~=6.0.0
python-dotenv~=1.1.0
webdriver-manager~=4.0.2
chromedriver-autoinstaller~=0.6.4